Se corrige error que obtenía el estado de los  switches y no de las luces.

+++ guided_q_learning/env/light_env.py
@@ -198,7 +198,7 @@ class LightEnv(gym.GoalEnv):
         self.steps = 0
         self.correct = []
         # self.state = np.zeros((self.num))
-        self.state = np.random.randint(2, size=self.num)
+        self.state = np.random.randint(2, size=self.num).astype(float)
         self.eps += 1
         # self.goal = self._sample_goal()
